+++ traktor-mod/D2/Api/ApiChannel.qml
@@ -1,6 +1,6 @@
 import CSI 1.0
 import QtQuick 2.0
-import "ApiClient.js" as ApiClient
+import "TrackIdApiClient.js" as ApiClient
 
 Item {
   property int       index:            1